fix to terminate the process that has ended
check if the process exists before using teminate() to stop the process.

+++ moviepy/audio/io/readers.py
@@ -1,6 +1,7 @@
 import os
 import subprocess as sp
 import warnings
+import psutil
 
 import numpy as np
 
@@ -249,10 +250,11 @@ class FFMPEG_AudioReader:
 
     def close(self):
         if self.proc:
-            self.proc.terminate()
-            self.proc.stdout.close()
-            self.proc.stderr.close()
-            self.proc.wait()
+            if psutil.pid_exists(self.proc.pid):
+                self.proc.terminate()
+                self.proc.stdout.close()
+                self.proc.stderr.close()
+                self.proc.wait()
             self.proc = None
 
     def __del__(self):
